IDEA-319986 Fix invisible root collapse bug in DefaultTreeLayoutCache
Add some tests while we're at it.

The bug was breaking AsyncTreeModelTest. In production code
collapsing the invisible root is pretty unlikely, as it leads to an empty
tree with no obvious way to expand it back, so nobody should provide
a way to collapse it to begin with.

(cherry picked from commit 5c7e41281d06693cc843046f1ee13a53c89af475)

GitOrigin-RevId: b46fd02ff8a66b785afdab5a1233e77436cbdb66