IDEA-319986 Fix re-expanding bug in DefaultTreeLayoutCache
Yet another bug not really visible in production. When a node
is collapsed, all its children are normally collapsed too.
But if it's collapsed through the API, then the children
remain expanded. This means that when the node is
re-expanded, its children should also be re-expanded
back to the previous state. Normally it isn't required
because they're previously collapsed by the tree UI,
but if we collapse just one node through the API
without explicitly collapsing its children, then this
bug becomes apparent.

It was causing the AsyncTreeModelTest.testChildren
to fail. Now it's all green.

(cherry picked from commit 801ddeed6f91243293567427dc1b4b2d41b02243)

GitOrigin-RevId: 0a7b8161b558c1be83bcb4eba787f7917c4f1815