IDEA-323451 Fix IDE Errors Dialog font size
In case of using a custom font size for the editor, it was incorrectly scaled.
So instead of just copying the font from the editor settings, use it to
derive a new font with the same size as the usual label font.
Also, don't try to scale it on creation, as labelFontSize is already scaled.

GitOrigin-RevId: 40a5e9bab963b3fed4fa1572ac145833aebf3e9d