IDEA-319986 Fix more resetRowSelection() bugs in DefaultTreeLayoutCache
Not sure if these are really bugs, but the safest bet is to mimic
the behavior of VariableHeightLayoutCache exactly, to avoid
any surprises from deep within the library.

Specifically, when a node is collapsed, the selection is only
reset if it was visible and the number of rows has actually changed.
This can only be the case if the invisible root was collapsed,
which doesn't make much sense anyway.

This has a nice side effect that we can always test our tests
by replacing the SUT with a VariableHeightLayoutCache
and see if they pass.

(cherry picked from commit 9e7f3b60282812f828104226b48d13bf4e79f6f7)

GitOrigin-RevId: 613b50cf280ae8769b7e8fa4bb783d2e58d5764d