RDCT-315 Support showing Hint under gutter mark
Introducing new API to HintManagerImpl: showGutterHint
In comparison to showEditorHint, this method doesn't need exact point where it should be shown. Instead, HintManager calculates it based on passed line and horizontal offset.

The reason to add the new API is that in RD/CWM scenario we can't get the correct position (due to differences in location/sizes/UI hierarchy), so it is better to pass more simple data which would stay the same for host and client.

GitOrigin-RevId: 2a9196258b8261af4e60f471afd8995b27c5dc1f