[build scripts] change the version format for plugins built from intellij sources (IDEA-312812)
We need to ensure that a version of a plugin built from sources is greater than any fixed version uploaded to Marketplace, so 'SNAPSHOT' cannot be the second component of the version since it's considered to be less than any number. The date number is big enough to ensure that the second component of any version from Marketplace will be less than it.

GitOrigin-RevId: 0dc59597f176e24524688b9bc866b1551fc21440