[debugger][kotlin] IDEA-323489 Fix evaluation of expressions with type arguments
Now the evaluator erases type arguments to their upper bound, in addition to already supported type parameters.

Generic types in a generated function may appear in the return type (when the expression itself has a generic type), or when accessing members of a generic class (so the generic class instance is passed to the function as an argument).

Also fixes EA-847089 IAE: BaseKotlinMangleComputer.mangleTypeParameterReference (some cases at least).

Merge-request: IJ-MR-109504
Merged-by: Alexander Kuznetsov <Aleksander.Kuznetsov@jetbrains.com>

GitOrigin-RevId: 1f8332f53e31bc2190fcf0ecea7f864daa129605