[debugger] IDEA-322970 Fix freeze on breakpoint in inline function
Search for inline function callers is done in a non-blocking read action, that must be cancellable. In this commit, we make sure that the outer progress indicator is not replaced by a new one during the call. Cancellable read action lets a concurrent write action to execute successfully, so the UI is not freezed anymore.

IJ-MR-109631

(cherry picked from commit 9cbf218ee3955b4bf0593daabaa4e0abd9839525)

GitOrigin-RevId: bd094ea726da1dbc836309a32d6c862663d8e5bf