PY-59132 Ask for user confirmation when installing packages with a quickfix
Namely, if a package doesn't belong to the list of "well-known" packages, which
we maintain for the Packaging Tool Window, such as Django, upon installing it
with a quickfix (either "Install package" or "Install and import package")
we ask for an explicit confirmation, mentioning that a package can execute
potentially malicious code while being installed. This extra step should allow
a user to double-check a package name, making sure that it's not a spoofing
variant. This confirmation dialog can be disabled IDE-wide.

IJ-MR-109660

GitOrigin-RevId: 53e08ef558be0580d5978d0b8fd40e9f7722b1a1