[rdct] use the same layout of per-process config/system directories even on the first run
Before, when JetBrains Client is started for the first time on a new machine and needs to import config, it uses the standard config directory. In that case, synchronization of settings with other processes works differently, it writes to the configuration files directly, not via StreamProvider. It complicates testing and may lead to I/O errors since direct access doesn't lock files (RDCT-353). Now an ability to import config to a custom directory is added to ConfigImportHelper, and ThinClientPathCustomizer always uses the same layout of config/system directories, all processes are equitable and always use StreamProvider to read and write shared config files.

ConfigImportHelper was changed to always use the passed config directory instead of using the default PathManager.getConfigDir value. ConfigImportLicenseTest don't use in-memory FS anymore, because it doesn't work properly when files are copied in backupCurrentConfigToTempAndDelete.

GitOrigin-RevId: b8ab3bce78c5efe2cf8fe15f465108042b76b6bb