IDEA-323529 Relax requirements for restored tool window bounds
It used to call WindowManager.isInsideScreenBounds, which is broken
for multi-monitor setups with scaling, at least on Windows, because
the coordinate space is not continuous. E.g. an FHD monitor scaled to 125%
to the left of the main one at x=0 could have x=-1920 and width=1536,
creating a gap of 384px between the monitors that isn't really visible.
If a window was spanning multiple monitors when it was closed,
then these bounds would be considered invalid when the window is opened
the next time.

Fix by resorting to ScreenUtil.isVisible instead. There's a catch, however:
it only checks that some 10% is visible. That's better than nothing,
but for windows with headers specifically it's not always enough, as if
the header is offscreen, it could become a huge problem. So add some
isVisible checks for the top-left and top-right corners too. There's another
catch here: such checks may fail if the monitors are aligned vertically
with a coordinate gap between them, and most of the window is located
on the bottom screen. If that's the case, the header will be considered
invisible. It isn't easy to fix without a proper API, but it's not a big deal
since it's a rather exotic case and it's a false negative: in such a case,
the window will be re-shown at the default location, it won't be invisible
or cut off in an annoying way.

GitOrigin-RevId: 8321c2a2593b43e19b445197fd34bc2d2896e0ca