IDEA-318081 Use NamedBorder for banners to update it on LaF change
We really need some smart borders for this to avoid overriding updateUI()
every time. Here it was already overridden, so it's not a big deal, but still
lots of boilerplate code that could be avoided. But it'll have to do for now.

(cherry picked from commit 874c6f5031188d1f5976856a0ea5eaf4f3972804)

IJ-CR-109764

GitOrigin-RevId: f9d98841766ec4944987e7f0499b2cbd07e336f4