[gradle]: Rebuild buildSrc if changes only in buildSrc module
BuildSrc modules are not rebuilt when delegation to gradle is enabled,
because gradle would build it automatically. But when there are only
changes in the buildSrc module and nothing else is build, buildSrc is
still skipped. This change adds a check for that siutation and then
executes an dummy task in the root project created by an init script,
to let gradle build the buildSrc.

Issue: IDEA-280826