ML in SE: Fix "blinking" of results in SE (don't score typo fix by ML) (IDEA-323795)
The blinking is a side effect of resorting a list
shortly after typo appears at the top. Almost
instantly, it gets ranked by a ML model, and
then the list gets sorted based by the new
priority, causing the item to appear at the very
bottom of the list.

This happens very quickly, and looks like
the list of results is blinking

GitOrigin-RevId: a16506cbece18d87122c9c878adab4657ea514fa