[kotlin] J2K: drop AddOperatorModifierInspection post-processing
This is an information-level inspection. We can't automatically determine
if it is appropriate to add 'operator' modifier to any particular function.

Actually, the affected test `kt-19340.kt` is a perfect illustration of this:
the function `set` doesn't have the semantics of an operator function in Kotlin's sense.
It uses its two parameters differently: indexed access doesn't make sense in this case.

GitOrigin-RevId: 41c252f3b7e64ec768265166525c41bc3eb13ef7