IDEA-323679 Hide/show the stripe button when restoring tool window layout
The setToolWindowAnchorImpl() method already contained the code
to show/hide the stripe button is needed, but it wasn't invoked when
isShowStripeButton has changed. This caused tool windows with
the stripe button currently hidden not to be restored when restoring
a layout previously saved with that button visible.

Fix by simply adding another condition to invoke setToolWindowAnchorImpl().

GitOrigin-RevId: ffaab9591249e4a44835d03e4b4f124592bc6538