[kotlin] OptIn support in build.gradle.kts
Since scripts (in particular, build.gradle.kts) usually uses a nested structure, the logic for finding candidates for setting annotations in scripts has been changed: will be suggested setting an annotation on current the statement, as well as on all parents to which this applies. Also, SAM interfaces are supported.
For files with the .kt extension, the logic is preserved: it will be proposed to add an annotation on the first non-local parent and the containing class, if it exists.

^KTIJ-25375

GitOrigin-RevId: 6034b4847ce1f73a2cb89cfc268096e43e6729ca