IDEA-323550 Update visibility of tool window toolbars as they are added
Well, it was a tricky one. I'm not really happy with the way it turned out,
but I can't think of a better solution. The problem is that toolbars are created
and added pretty much anywhere, including plugins. There is no single API
for adding a toolbar to a toolwindow, and it doesn't seem there should be one
because different tool windows have different content, and it's created
in different ways, which makes perfect sense.

Whenever the ToggleToolbarAction is performed, it basically traverses
the whole UI hierarchy, looking for toolbars and updating their visibility
in the process. However, this didn't account for the initial visibility,
hence this bug. We can't just update visibility of toolbars on a tool
window creation, however, because at that moment its content may
not be fully initialized yet. Even worse, it could be (and often is) dynamic,
with new tabs with their own toolbars created as needed.

Seems the only way is to update visibility of toolbars as soon as
they are actually added. To do that, we hijack their addNotify()
method in a rather ugly way, sending a message through the message
bus on the ActionManagerListener topic. Then we can check whether
it was added to our tool window and update visibility if needed.

GitOrigin-RevId: 96202a6841461d2b91d36258afbc9c4a34173272