IDEA-323645 Fix app icon scaling
Another scaling mess. After the KDE fix (IDEA-320504),
the size of the app icon was fixed to 16. This is
a completely unscaled value.

It would then be scaled by the pixScale factor
to render the SVG image, and that was fine.
The problem is with JRE HiDPI, when the same
size would be used to determine the size of the icon
in the user coordinate space. This is, of course,
incorrect, and that's why it was so small on Windows,
where the JRE HiDPI mode is the default.

Fix by introducing a whole mess of scaling factors
into loadAppIconImage():

- pixScale is used to find the SVG;
- userScale is used to calculate the size
in the user coordinate space to pass to loadWithSizes(),
which then uses it to properly construct a HiDPI-aware
image in case the JRE HiDPI mode is used;
- sysScale is used to pass to loadWithSizes(),
which then uses it to calculate the real size
(to render the SVG) and, again, to create a HiDPI-aware
image.

GitOrigin-RevId: 94d1953dbed8c419ec8616b098967545486b5ab4