Refactor refresh requests flow processing in SaveAndSyncHandlerImpl.kt
The commit moves the logic inside doScheduledRefresh() directly to the collection of refreshRequests within the coroutine. This eliminates doScheduledRefresh() function and simplifies the code structure. The function was only called from one place and its logic is more appropriately co-located with the triggering event. In addition, minor grammar correction was made in a comment

GitOrigin-RevId: 552f389ff158826cc988ad7ca1e3e5816010be81