CWM-6550 Use AnchoredPoint to calculate the location on a component with empty size
In RD/CWM the size and location of most components is 0 as components aren't actually showing.

If we create RelativePoint using those coordinates, we will end up showing a popup in the wrong location. As this happens in CWM-6550.

We can't operate sizes and location of a component on a host, but we can use some abstract values like anchors.
So, I'm introducing AnchoredPoint (inheritor of RelativePoint).
Instead of calculating relative point as (x + width, y + height) we just say that AnchoredPoint is a BOTTOM_RIGHT point on a component.

And then we pass this anchor to a client which will be able to calculate the proper location.

GitOrigin-RevId: 838e64950e63f882750e105e4ca4f985eea7e05b