[settings] synchronize settings of storages with RoamingType.DISABLED in process-per-connection mode
Before, StreamProvider wasn't called for a storage with RoamingType.DISABLED, and this caused problems with persisting such settings in JetBrains Client when separate processes are used for different connections (CWM-8334, RDCT-366).
Now StreamProvider is used for all types of storages, and the existing implementations were changed to skip storages with RoamingType.DISABLED.

In order to keep the old behavior, when a storage which type was changed to DISABLED is deleted from StreamProvider, a new method 'deleteIfObsolete' is added and implemented in IcsManager and CloudConfigStreamProvider (SettingsSyncIdeMediatorImpl didn't do that before). This is a temporary solution to minimize the amount of code changes, it would be better to implement this logic in 'StreamProvider::write' method, but this requires getting rid of 'isAvailable' calls, so this should be done later.

(cherry picked from commit 2c660d18ca6ba0c66aed9c01d68bbae0bc323edb)

IJ-CR-110355

GitOrigin-RevId: 41508a0390e6c6574de4746eba8b07c282e25843