Prevent flaky shared indexes build failures due to thread pool starvation
Look at the snippet below:
```
val initializationJob = scope.launch {
  bundlePaths.map { bundlePath ->
    launch(Dispatchers.IO) {
      print("hello world")
    }
  }
}
```

The pitfall is that initially it starts a coroutine in the default thread pool, which has a hard limit of concurrently launched tasks. Only when there is a free thread, the coroutine launches and schedules new coroutines in another, unlimited thread pool from Dispatchers.IO.

This becomes more dangerous when this asynchronous approach is combined with blocking approach. Here, `ProgressIndicatorUtils.awaitWithCheckCanceled(initializationJob.asCompletableFuture())` can be executed in a thread from the very same limited thread pool, increasing the probability of thread starvation.

GitOrigin-RevId: b9e7f765547ca8199ab32b9df35b7b6b60712166