avoid calling actions update during the main menu calculation
As the main menu now reuses the ActionUpdater logic, we need to restore the always-visible semantic that was lost in transition. Main menu subgroups do not need to check whether they have any enabled children to be visible. The check is quite expensive, e.g. `PresentableActionHandlerBasedAction.update` is called.

We introduce ALWAYS_VISIBLE property to add "always-visible" flag to existing action groups.

GitOrigin-RevId: b7fd5f2b616682786037910b09165c1e98dc25f3