IDEA-324341 Implement auto scroll to folding regions in Structure
The problem with folding regions is that they aren't PSI elements,
and are added to the Structure View tree independently.
Therefore, when TextEditorBasedStructureViewModel.getCurrentEditorElement()
is searching for the current element, it can't possibly find a folding region
and finds the containing element instead.

Then, when the tree is traversed to look for the element, visiting
stops right at that containing element because canRepresent()
rightfully returns true.

Fix by considering all matching elements to be possible ancestors
of the element we're looking for, and at the same time checking
not just for the element, but also for the text editor offset, if any.
So instead of returning INTERRUPT, we continue visiting children
and if any folding regions are found, we check whether the current
region contains the editor offset, considering them a valid match
in this case.

This has a side effect of possible performance degradation since
visiting no longer stops when a matching node is found. Instead,
we always rely on finding the longest tree path, and the resulting
promise always receives a null result. Since visiting is performed
on a BGT now, that may not be a huge deal, though.

GitOrigin-RevId: 751f43613ececd85da2435c5281115aa45ea2af3