Allow testing syntax highlighting via answer as string
In this commit we extract the serialization logic out of
`testFileSyntaxHighlighting` and add an additional method
with the same name, which accepts the expected serialized
highlighting directly as a string. This way a separate file
with answers is not needed anymore.

(cherry picked from commit 195ca9411d01f70fe44ae8ac3dd206e9476202b6)

IJ-CR-111727

GitOrigin-RevId: 4a875d95add719bfe8fa13fce4ae04d8a93c9011