Allow testing syntax highlighting via answer as string
In this commit we extract the serialization logic out of
`testFileSyntaxHighlighting` and add an additional method
with the same name, which accepts the expected serialized
highlighting directly as a string. This way a separate file
with answers is not needed anymore.

(cherry picked from commit db41ecc99c8c5a8cdba974cec4984e595dc84849)

IJ-MR-110574

GitOrigin-RevId: 6b0f2bf59d867a0ffb412209e1aa6b2b8372dfdc