IDEA-315879 Add assertIsEdt() to AsyncTreeModel.dispose()
Disposing on a non-EDT is not thread safe. The model may be in the process
of processing an update, and it may be running a listener, notifying it about
that update. Those listeners have no other way of retrieving the updated nodes
but to fetch them one-by-one by calling getChild(). Once the model is disposed,
that method immediately starts returning nulls, which breaks the client.

There is no real way around it as the Swing API was designed strictly for
single-threaded use. It is in theory possible to postpone disposal until all
listeners have finished, but it may mask other problems caused by asynchronous
disposal, and not only of the model itself, but of its parent Disposable,
which may be a single-threaded UI component too.

So let's try first to ensure that the model is used (and disposed) as intended,
strictly in the EDT.

GitOrigin-RevId: 527a66edae3f7819edc15c81df5543baf143604c