IDEA-315879 Dispose AsyncTreeModel in EDT in BuildViewTestFixture
Now that dispose() is only allowed in the EDT, fix the tests accordingly.

It may make more sense to only dispose the build view on the EDT,
not the entire fixture, but it's more tricky, and it seems to work
fine this way, so we can get back to it if it causes any problems.

(cherry picked from commit b7442e187918c48e1caee0327a5b6b41d37b5ea9)

IJ-CR-110639

GitOrigin-RevId: 876fd665f287f1d0ca0d9e2178f30460e22dc7b1