IDEA-324855 Don't perform "Is editor visible?" check for Undocked tool windows
Caused by over-aggressive sanity checks added for IDEA-319836:
when a docked tool window is resized, we don't need to remember its size
if it's in a splitter without a visible editor, as that screws up the size.

However, when it's Undocked, it's not in the splitter, and that sanity
check is not necessary, we can safely return true and skip it.

GitOrigin-RevId: c13aab913179b4cd6e13ece58dbf374baf30c5dc