IDEA-324832 Hide stripe buttons on restoring layout if they were hidden
Strangely enough, doSetAnchor() was always creating stripe buttons,
regardless of the isShowStripeButton value. Fix that, but as a safety
check show the stripe button anyway if the window is visible.
The combination of isShowStripeButton = false and isVisible = true
shouldn't happen anyway, but just in case the layout is corrupted,
we don't want to show a tool window without a stripe button.

GitOrigin-RevId: 292f374cadc39e1dd00168c6aec4c78f81c30878