Prevent error in UI scaling prior to application start
Adjusted the conditions in `JBUIScale.kt` to prevent the triggering of a precompute error when the `uiDefaults` are null and the application hasn't started yet. This is to ensure smooth initialization of the scaling features under defined circumstances.

(cherry picked from commit 22fcb7e8d9f55d6ecc51437604ae4ae8ab949681)

IJ-CR-110964

GitOrigin-RevId: 1e580d9a5167bd56b0753ec53fb7f450569917f2