[rdct] fix installation of non-dynamic plugins for "process per connection" mode (RDCT-381)
Before, plugins which require restart were downloaded to idea.system.path/plugins directory, and commands to install them were written to action.script file in the same directory. But idea.system.path was cleared when the client process was restarted, so plugins weren't installed. Now a special method PathManager.getStartupScriptDir is introduced with an ability to override it via PathCustomizer. ThinClientPathCustomizer sets it to a directory which won't be cleared after restart.

GitOrigin-RevId: 3c7850b74168c5de9bf796fbd6dcc9552b1d8f7e