Refactor MarkdownHtmlPanelProvider for cleaner code & improved readability
In this commit, unnecessary import statements were removed and annotations rearranged to follow coding conventions - increasing the legibility of the code. The method getProviders() was modified to return a List instead of an array providing better flexibility and standard usage. Furthermore, a number of methods were reformatted for consistency, and the ProviderInfo and AvailabilityInfo classes were refactored for improved readability.

GitOrigin-RevId: 7c898a2225a970dee23fb5ccef995e6140162367