merge fast-track phase into ordinary update
If requested, EDT will wait for fast-track timeout for the ordinary session to complete. To achieve that:

1. Custom later invocator (now executor) switches from blocking queue default "invokeLater" on-the-fly
2. Data-context is fully-functional. "Freezing" is no more needed
3. "Fast" and "exclusive" executors are no more bound to 1 thread, and grab the new one when needed

(3) is ok as long as fast-track is always a user-triggered action. There shall be about one thread at any time. Except when there are several toolbars that are about to appear on screen for the first time. Such fast-tracks are limited to one thread at a time.

GitOrigin-RevId: bbe76f48cdd49b58c06bc3b7948f6196bc582db9