IDEA-324341 Fix erroneous closest PSI element selection in structure view
Well, it's an odd one. FileStructurePopup contains some logic that
causes the nearest to the found PSI element to be selected,
not the found one. This was apparently done to fix DBE-2249,
but for some reason only in FileStructurePopup and not in
the StructureViewComponent.

This logic was broken by the fix of folding region selection,
which goes deeper even if the element is found, eventually
reporting back the element with the longest tree path.

However, that nearest-PSI-element logic only kicks in
when the longest path match is used. In other words,
if an exact match is found, then the found element is
selected. If the longest-path match is found, then
the nearest element is selected. Since after the fix
the visitor always returns the longest-path match,
this logic now kicks in when it shouldn't.

Fix by storing the kind of match (exact or longest path)
and analyzing it after the search is complete. So
we still go deeper and check every node, but remember
what kind of match it was.

It looks like the same logic should be applied in
StructureViewComponent as well, and duplicate code
should probably be removed. But that's a different task.

GitOrigin-RevId: f329f3f6f5217eca7bf7673dff76e273e0648118