Extract StructureViewSelectVisitorState
That stage / deepestPath thing became too ugly.
Extract it into a separate class.

Implementation note: the condition effectively
changed from (count == 0 || count < pathCount)
to just (count < pathCount) because pathCount
must be positive anyway. If it isn't, it's either
a bug or a weird empty tree that allows visiting
the non-existing root, which doesn't make
much sense, and we can't select it anyway.

GitOrigin-RevId: ac016614dcc1ca03103fc667a6c5b3e2759b4b25