IFT-470 Fix absence of sample file in onboarding lessons
The problem occurs only after rewriting the whole learning project in `ProjectUtils.importOrOpenProject` in case of major IDE version change or manual removal of the learning project from disk.
There was no VFS refresh after the rewriting, so VFS said that sample file exists, but there was no file in the real file system, so we didn't create it in the lesson script.
As a solution, I added the refresh before each project lesson start. I think it will be more reliable.

GitOrigin-RevId: 1b4941f53feaaab12c06ef6d672546ebb43d194d