EA-883919 Don't try to set window type in AbstractPopup if the window is displayable
Couldn't reproduce it locally, but it's possible that the window previously hidden
was reused later while still having a peer. The simplest way to fix it is to
simply check if it's displayable before showing it. This, of course, has
a side effect of not applying the new window type if it's changed since
the popup was shown the last time, but it shouldn't happen normally,
as if we're reusing the same popup, the type should stay the same as well.

GitOrigin-RevId: 5821ee3c3635daa702f17fa24849ae3c685146da