IDEA-325102 Use a flow for select in project view to order requests
Previously selection was handled by a separately launched coroutines,
which didn't guarantee the order in which elements were eventually
selected. This sometimes caused a later selection to be replaced
by an earlier one.

Fix by moving everything into a single coroutine and using a flow
to send selection requests there. No need for a replay buffer here
because the coroutine always sits there collecting requests anyway,
and the buffer size is 1 because we don't want to suspend the caller
(it's not even possible, as callers are outside of coroutines)
and we don't need anything but the latest selection request.

GitOrigin-RevId: 50783471fd717896fba84c494d222ce9fc046a15