IDEA-325296 Don't auto-expand files in project view
By default, when a tree node is expanded, and it has only one child,
it's also expanded. However, for the Project View it makes little sense
if that child is not a directory, but, for example, a class (with the Show Members
option on). A class may contain many members and expanding it
is more often annoying than useful.

If a scenario comes up when it's actually useful, we can always make
it optional, but for now let's just change the default.

Implement by installing an auto-expand filter on the project view
tree that prohibits auto-expand if a node has an file associated with it
and that's not a directory.

GitOrigin-RevId: cf3213d32568e2091bcdf5c025309b0744fa372f