Refactor getTracer method to use Scope instead of separate parameters
The getTracer method in the TelemetryManager and TelemetryDefaultManager classes has been refactored to use a Scope object as a parameter instead of separate scopeName and verbose parameters. This change simplifies the method signature and encapsulates the details within the Scope object. The Scope object's toString method is used to obtain the scope name. The verbose property of Scope object is used instead of a separate parameter. Also, the usage of RDCT_INSTRUMENTER_NAME constant in TelemetryDataUtil class has been replaced with a Scope object. CodeVision has been made verbose by default.

GitOrigin-RevId: 9619ab169250bc72328438cd2898480c248a4fe1