[kotlin] k2: allow going deep from java overriding search when first layer doesn't provide any result
^ KTIJ-16021

The LC search was prevented from performing searches initiated from `Java overriding search` because a non-LC search performs an in-depth, level-by-level search through inheritors itself. However, if this non-LC search fails to find results, because inheritor doesn't have an override function, the LC search should be allowed to attempt the search. This ensures that indirect Java inheritors are not missed.

GitOrigin-RevId: e650d14aeef83de18e517df75a8f61879a0af62f