[mod-commands] Basic support of editing kotlin light classes
The class object we have in hands is an ultra light class. Its containing file is a fake file (org.jetbrains.kotlin.asJava.elements.FakeFileForLightClass) which cannot copy itself properly. Even if it could, we would probably have troubles finding the copied ultra light class.

Instead, we try to solve this using navigation element (which is the original kotlin file) and PsiClassOwner interface, to materialize the proper copy

GitOrigin-RevId: 9172ba3d821ac82babbc94da3d091c08ec28a6cc