IDEA-325212 Don't close the popup if a modal dialog opened from it is displayed
If a modal dialog is opened from a popup, then closing the popup because
it lost focus (because of the dialog receiving it) or because the user clicked
in the dialog (and therefore outside the popup) is wrong, as the usual contract
of a modal dialog implies that only the dialog should process all user inputs,
the expectation is that nothing outside of it changes because of those inputs.

Moreover, when the dialog is opened to perform some actions on that popup,
like Paste from History in the referenced ticket, it's obvious that the popup
should remain open while the dialog is shown and right after it's closed.

Fix by checking the current modality state in AbstractPopup.canClose().
If the current modality state is more specific than the one that was the current
when the popup was shown, then the popup is not closed. Moreover,
in that case even the close callback is not invoked, as some clients
use that callback as a close notification, in the form of "do stuff, return true",
to hide/close/dispose of some stuff used by the popup, and we definitely
don't want to do any of that if the popup is not going to be closed after all.

GitOrigin-RevId: ceb748d0104054b858ab3506cba69ba614d62ce7