IDEA-325562 Fix collapse icon for the new run configuration popup
Use the same top/bottom insets for the new UI as used for other
popup headers. TitlePanel (that shows the title in the popup header)
uses the same insets, so everything is neat and in line.

Implementation note: inset values have to be un-scaled first,
as they're represented by JBInsets, and if we just used them
for the border, then the border itself will scale them again,
screwing up insets completely, which is very noticeable
when using IDE Zoom: the icon is shifted badly then.

Leave the old UI alone for now. It's a bit off there as well, but
not as horrible, and it's not clear where to get the correct values
for it.

GitOrigin-RevId: d946d9fa90c8a8b3f686540377c42838bdf9450d