[Android Debugging] Fix breakpoint gutter markers.
Currently, KotlinLineBreakpoint assume that source debug extension
is not available on dex VMs. Therefore, it assumes that it cannot
determine if a loaded class contains a given position. That leads
to a lot of breakpoint gutter markers that are greyed out because
we first say incorrectly that a position might be in a class,
however, we then later see that it is not and invalidate the
breakpoint. When the correct class is actually loaded, the gutter
marker goes back to a normal red circle as the breakpoint is
correctly set.

See: https://issuetracker.google.com/issues/291060622

Starting Android O, source debug extension is available on dex VMs
and therefore the code that is skipped on Android works in the
vast majority of debugging cases these days. Even if it doesn't
the breakpoint will not be set in any case if we cannot find the
position in the class that was loaded. So, we just remove the
dex bailout as it does no good.

The Kotlin position manager similarly has code that tries to
make breakpoint setting for inline functions work even when
we do not have source debug extension. I left this code intact
except that I now check whether we can get the source debug
extension to avoid the approximation when we can get the
source debug extension. The breakpoints set when there is no
source debug extension available do not actually work: you will
stop at the end of the file the code is inline into, the
breakpoint will be marked invalid and will therefore not trigger
again until you remove it an readd it. I'm not sure how
useful that actually is and we could consider just removing
it.

(cherry picked from commit 276d19b6d8501fd444493b9d21c95b03ce306cef)

Bug: 236897529
Test: Manual
Change-Id: I6b6eb8cab72558a509c6d6ebc8c1127716087ab5