Deprecate SelectInContext.selectIn
It has only one external usage and a few internal usages that will
be fixed eventually.

The reason for deprecating it is simple: canSelect() and selectIn()
should be called on different threads (BGT and EDT respectively)
because canSelect() often involves PSI slow ops, and selectIn()
usually involves the actual selection, that is, UI update.

GitOrigin-RevId: ec6ab827968830582e80ce83d1fd8037c979a956