IDEA-322959 Split context.selectIn() between BGT and EDT in FileNavigatorImpl
To get rid of the EDT slow op, it had to be split between two calls,
each executing in its own thread. Couldn't reuse SelectInProjectViewImpl
because it's from lang-impl, so created yet another light project
service just to get a coroutine scope.

Now there's a slow op in ProjectViewSelectInGroupTarget.selectIn
for the same reason, that's the next step.

GitOrigin-RevId: 6a40c327fc6fe878ad95e0bd39abea111fe90b74