[KTIJ-26322] Find SourceSet DataNodes using `GradleProjectResolver.RESOLVED_SOURCE_SETS`...
to optimise import performance for larger projects (such as Ktor)

The function
    `fun DataNode<GradleSourceSetData>.addDependency(dependency: IdeaKotlinSourceDependency)`

which is setting up source dependencies between two modules
requires to resolve the module associated with the dependency.
It therefore tried to go through the whole project, trying
to find the DataNode<GradleSourceSetData> which has the same coordinates
as the passed dependency.

This function gets called

For 'source dependency' in every SourceSet and then traverses
all data nodes to find the dependency source set

(therefore the complexity is at least quadratic in the number of SourceSets)

The function called 'findFirstRecursively' was then also allocating
a queue, which seemingly was slow as well.

However, there is already a map maintained called 'GradleProjectResolver.RESOLVED_SOURCE_SETS'
which makes looking up the data nodes O(1)

^KTIJ-26322 Verification Pending

GitOrigin-RevId: 1eaa8472645d48416ddc71ecac4396b4dd0b4b74