IDEA-323344 Add `Start Quick Tour` button to the Meet New UI tool window
Unable directly access NewUiOnboarding from the MeetNewUiToolWindow, because onboarding is not in the platform, it is a product module. So implemeted it using the extension. Probably it is worth to move MeetNewUiToolWindow to onboarding module, but it can be possible when all IDE's will enable the onboarding.

(cherry picked from commit 084e546139d037a52bc549722db52613d37d38da)

GitOrigin-RevId: fc03899ea1d04362b067a77bbb628f1d028765c0