Support Linchek stack traces: IDEA-324320
The Linchek tool mentions so-called events before `at` word, which indicate which method was called with which parameter.

The normal exception parser doesn't work here, because the method call matches the `(<digit>)` pattern, and it confuses the exception parser which thinks that it is the left-most appearance of a file with line. To avoid this, we specifically search for the `(<digit>)` after the "at" word.

GitOrigin-RevId: cbc7c2523fceaa30d819d1a6c982c99dd3c0c5fb