[Kotlin] MPP: remove the handling of the HMPP flags from the test infra
HMPP is enabled by default since the 1.6.x, and all tests that we have right now using the KGP >= 1.7.0
I.e. isHmppEnabledByDefault is true all the time, so significant part of the code just removed.
The only place where test behaviour actually was changed is in NewMultiplatformKaptProjectImportingTest.
For some reason the tests there were setting the `{{ disable_hmpp_flags }}`, but they are okay after the HMPP enabling

#KT-59304 In progress

GitOrigin-RevId: 119035dc1e12c9a1b725aa72abf1881281803658