IDEA-325896 Log popup bounds calculation in LookupUi
No logic changed for now, only logging, although
it's already clear that the myPositionedAbove calculation
line contains at least one bug and one redundant condition.

Some code was refactored to avoid calculating the same things
twice, not because of performance reasons, but to be absolutely
sure that these values don't change between calls (shouldn't
happen, but with logging it's best to be absolutely sure).

GitOrigin-RevId: ccbbb7a389c0e747e1ab7e86bd707b761c2686fb