IDEA-326126 Refactor extended info check for search everywhere feature
This commit refactors the "search.everywhere.footer.extended.info" check in the search everywhere feature. Previously, the check was performed directly against the Registry. Now, it's abstracted into a new method "isExtendedInfoEnabled()" in the SearchEverywhereUI class. Abstraction improves code readability and ensures that if more conditions are added to enable extended info, changes will only be needed in one place.

GitOrigin-RevId: bcc68543772e2110ae580a81020c7c482d13b018