Refactored application initialization workflow for better thread management
Reworked the code that handles the initialization of various services during application startup. This was done to improve the management and readability of coroutines used on startup. Additional changes include removal of redundant or unused code and class implementations, and fixing typographical errors in naming conventions.

This adjustment aims to ensure that Initialization tasks are effectively managed and scheduled, leading to potentially faster startup times and simplified debugging process during application startup.

GitOrigin-RevId: f674cf8771d2be1ebf59c166f439aa740af8fff9